+++ crates/flipperzero/src/furi/mod.rs
@@ -15,4 +15,4 @@ use flipperzero_sys as sys;
 /// Furi Result type.
 pub type Result<T> = core::result::Result<T, Error>;
 /// Furi Error type.
-pub type Error = sys::furi::Status;
+pub type Error = sys::furi::Error;